requirements.txt
cryptography
pillow
requests
SQLAlchemy
customtkinter